% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{internals}
\alias{internals}
\alias{current_popsize}
\alias{get_roottime}
\alias{get_tiptimes}
\alias{get_discrepancies}
\alias{drop_node}
\alias{nwk_recur}
\alias{write_nwk}
\title{internals}
\usage{
current_popsize(object)

get_roottime(object, prune = FALSE, init.include = FALSE)

get_tiptimes(object, prune = FALSE)

get_discrepancies(tr1, tr2)

drop_node(phy, nodes)

nwk_recur(p, df)

write_nwk(df)
}
\arguments{
\item{object}{\code{gpsim} object}

\item{prune}{logical}

\item{init.include}{logical}

\item{tr1}{character, newick}

\item{tr2}{character, newick}

\item{phy}{\code{phylo} object}

\item{nodes}{a character vector}

\item{p}{character}

\item{df}{tree data frame}
}
\value{
an integer

a numeric value

a numeric vector

a character vector

a \sQuote{phylo} object
}
\description{
internals
}
\keyword{internals}
