% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbdpwr.R
\name{lbdpwr}
\alias{lbdpwr}
\alias{LBDPwr}
\alias{runLBDPwr}
\alias{continueLBDPwr}
\title{Linear birth-death-sampling model with reassortment}
\usage{
runLBDPwr(
  time,
  t0 = 0,
  lambda = 2,
  mu = 1,
  psi = 1,
  rhoA = 0,
  rhoB = 0,
  n0 = 5
)

continueLBDPwr(
  object,
  time,
  lambda = NA,
  mu = NA,
  psi = NA,
  rhoA = NA,
  rhoB = NA
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{lambda}{per capita birth rate}

\item{mu}{per capita recovery rate.}

\item{psi}{per capita sampling rate.}

\item{rhoA}{per capita reassortment rate for segment A}

\item{rhoB}{per capita reassortment rate for segment B}

\item{n0}{initial population size}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}
}
\value{
An object of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{LBDPwr}.
}
\description{
The genealogy process induced by a simple linear birth-death process
with two segments reassorted
}
\examples{
simulate("LBDPwr",time=4) |> plot(points=TRUE)

simulate("LBDPwr",lambda=2,mu=1,psi=3,rhoA=.5,rhoB=0,n0=1,time=1) |>
  simulate(time=10,lambda=1) |>
  plot(points=TRUE)

# a pure birth process
simulate("LBDPwr",lambda=1,mu=0,psi=3,rhoA=1,rhoB=0,n0=1,time=4) |>
  plot(points=TRUE)
}
\seealso{
Other Genealogy processes: 
\code{\link{simulate}()}
}
\concept{Genealogy processes}
