% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.gpsim}
\alias{treeplot}
\title{Fancy tree plotter}
\usage{
\method{plot}{gpsim}(
  x,
  ...,
  time,
  t0,
  prune = TRUE,
  obscure = TRUE,
  hide = FALSE,
  compact = TRUE,
  retimes = FALSE
)

treeplot(
  tree,
  time = NULL,
  t0 = 0,
  ladderize = TRUE,
  points = FALSE,
  ...,
  palette = scales::hue_pal(l = 30, h = c(220, 580))
)
}
\arguments{
\item{x}{object of class \sQuote{gpsim}}

\item{...}{passed to \code{\link{treeplot}}}

\item{time}{numeric; time of the genealogy.}

\item{t0}{numeric; time of the root.}

\item{prune}{logical; prune the genealogy?}

\item{obscure}{logical; obscure the demes?}

\item{hide}{logical; hide reassortment events?}

\item{compact}{logical; return the tree in compact representation?}

\item{retimes}{logical; return the reassortment times?}

\item{tree}{character; tree representation in Newick format.}

\item{ladderize}{Ladderize?}

\item{points}{Show nodes and tips?}

\item{palette}{color palette for branches.
This can be furnished either as a function or a vector of colors.
If this is a function, it should take a single integer argument, the number of colors required.
If it is a vector, ....}
}
\value{
A printable \code{ggplot} object.
}
\description{
Plots a genealogical tree.
}
