% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_reassortment.R
\name{prune}
\alias{prune}
\alias{drop_before_root}
\alias{keep_shared_re}
\alias{drop_parent_re}
\alias{get_reversions}
\alias{compute_preRRtab}
\alias{refine_RRtab}
\alias{compute_RRtab}
\alias{Pruned_Re}
\title{prune the set of trees to dismiss invisible reassortment nodes}
\usage{
drop_before_root(init.nwk, seg.nwk)

keep_shared_re(init.nwk, seg.nwk)

drop_parent_re(init.nwk, seg.nwk)

get_reversions(init.nwk, seg.nwk)

compute_preRRtab(init.nwk, seg.nwk, incongr.init, incongr.seg, reversions)

refine_RRtab(init.nwk, seg.nwk, incongr.init, incongr.seg, RRtab)

compute_RRtab(init.nwk, seg.nwk)

Pruned_Re(object)
}
\arguments{
\item{init.nwk}{character; tree in newick format}

\item{seg.nwk}{character; tree in newick format}

\item{incongr.init}{character vector;}

\item{incongr.seg}{character vector;}

\item{reversions}{character vector;}

\item{RRtab}{data frame;}

\item{object}{\code{gpsim} object}

\item{nwktree}{character; tree in newick format}

\item{parent.numset}{numeric vector;}

\item{desc.numset}{numeric vector;}

\item{nodecol}{character;}

\item{incongr.nodes}{character vector;}

\item{reversed.nodes}{character vector;

\describe{
  \item{init.nwk}{the first-infection tree in newick format}
  \item{seg.nwk}{a segment tree in newick format}
  \item{incongr.init}{a character vector of incongruent coalescent nodes in the first-infection tree}
  \item{incongr.seg}{a character vector of incongruent coalescent nodes in the segment tree}
  \item{reversions}{a character vector of candidate reversed coalescent nodes}
  \item{RRtab}{a preliminary remove-and-rejoin table}
  \item{nwktree}{the tree string in newick format}
  \item{parent.numset}{a set of numbers indicating the candicate ancestors}
  \item{desc.numset}{a set of numbers indicating the candicate descendants}
  \item{nodecol}{the specified type/color (i.e. 'r' for samples, 'g' for branching nodes, '#Ha1' and '#Ha2' for reassortment nodes) of the requested nodes}
  \item{incongr.nodes}{the discording coalescent nodes between trees}
  \item{reversed.nodes}{the reversed coalescent nodes between trees}
}}
}
\value{
a character vector

a character vector

a character vector

a data frame

a data frame

a data frame

a data frame

a list of a set of trees and a data frame indicating the finalized remove-and-rejoin table
}
\description{
prune the set of trees to dismiss invisible reassortment nodes
}
