% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_reassortment.R
\name{Pruned_Re}
\alias{Pruned_Re}
\title{dismiss all invisible reassortments given the first-infection tree and segment trees}
\usage{
Pruned_Re(gp)
}
\description{
dismiss all invisible reassortments given the first-infection tree and segment trees
}
