% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_functions.R
\name{track}
\alias{track}
\alias{first.descendant}
\alias{last.descendant}
\alias{first.parent}
\alias{all.descendants}
\alias{inbetween.nodes}
\alias{notate.branches}
\alias{invisible.reassortments}
\alias{get.mrca}
\alias{track.theother}
\alias{fixed.reassortments}
\alias{tree.fixed}
\title{Track (a set of) components from a tree}
\usage{
first.descendant(parent.lab, nwktree, desc.numset, nodecol)

last.descendant(parent.lab, nwktree, desc.numset, nodecol)

first.parent(child.lab, nwktree, parent.numset, nodecol)

\method{all}{descendants}(parent.lab, nwktree, desc.col)

inbetween.nodes(parent.lab, child.lab, nwktree, nodecol)

notate.branches(coal.lab, nwktree)

invisible.reassortments(Re.lab, nwktree)

get.mrca(node.lab1, node.lab2, nwktree)

track.theother(nodelab, nwktree, nodecol)

fixed.reassortments(
  parent.lab,
  nwktree,
  incongr.nodes,
  reversed.nodes,
  fixed.renodes = NULL
)

tree.fixed(nwktree, incongr.nodes, reversed.nodes)
}
\arguments{
\item{parent.lab}{character; the branching node}

\item{nwktree}{character; tree in newick formate}

\item{desc.numset}{numeric vector;}

\item{nodecol}{character;}

\item{child.lab}{character;}

\item{parent.numset}{numeric vector;}

\item{coal.lab}{character;}

\item{Re.lab}{character;}

\item{node.lab1}{character;}

\item{node.lab2}{character;}

\item{incongr.nodes}{character vector;}

\item{reversed.nodes}{character vector;

\describe{
  \item{parent.lab}{the specified parent branching node}
  \item{child.lab}{the specified child node}
  \item{coal.lab}{the node label of the coalescent node}
  \item{Re.lab}{the node label of the reassortment node}
  \item{node.lab1}{a node label}
  \item{node.lab2}{another node label}
  \item{nwktree}{the tree string in newick format}
  \item{parent.numset}{a set of numbers indicating the candicate ancestors}
  \item{desc.numset}{a set of numbers indicating the candicate descendants}
  \item{nodecol}{the specified type/color (i.e. 'r' for samples, 'g' for branching nodes, '#Ha1' and '#Ha2' for reassortment nodes) of the requested nodes}
  \item{incongr.nodes}{the discording coalescent nodes between trees}
  \item{reversed.nodes}{the reversed coalescent nodes between trees}
}}
}
\value{
A string for the node/tip

A string for the node/tip

A string for the node/tip

A vector of characters

An vector of characters

logical

An string

A string

An vector of characters

An vector of characters
}
\description{
Track (a set of) components from a tree
}
