% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_functions.R
\name{track}
\alias{track}
\alias{track_first_desc}
\alias{track_last_desc}
\alias{track_first_parent}
\alias{track_descs}
\alias{track_between_nodes}
\alias{mark_removable_br}
\alias{track_Re_removed}
\alias{track_mrca}
\alias{track_theother}
\alias{get_desc_fixed_renodes}
\alias{get_tree_fixed_renodes}
\title{Track/get components from a tree}
\usage{
track_first_desc(parent.lab, nwktree, desc.numset, nodecol)

track_last_desc(parent.lab, nwktree, desc.numset, nodecol)

track_first_parent(child.lab, nwktree, parent.numset, nodecol)

track_descs(parent.lab, nwktree, desc.col)

track_between_nodes(parent.lab, child.lab, nwktree, nodecol)

mark_removable_br(coal.lab, nwktree)

track_Re_removed(Re.lab, nwktree)

track_mrca(node.lab1, node.lab2, nwktree)

track_theother(nodelab, nwktree, nodecol)

get_desc_fixed_renodes(
  parent.lab,
  nwktree,
  incongr.nodes,
  reversed.nodes,
  fixed.renodes = NULL
)

get_tree_fixed_renodes(nwktree, incongr.nodes, reversed.nodes)
}
\arguments{
\item{parent.lab}{character; the branching node}

\item{nwktree}{character; tree in newick format}

\item{desc.numset}{numeric vector;}

\item{nodecol}{character;}

\item{child.lab}{character;}

\item{parent.numset}{numeric vector;}

\item{coal.lab}{character;}

\item{Re.lab}{character;}

\item{node.lab1}{character;}

\item{node.lab2}{character;}

\item{incongr.nodes}{character vector;}

\item{reversed.nodes}{character vector;

\describe{
  \item{parent.lab}{the specified parent branching node}
  \item{child.lab}{the specified child node}
  \item{coal.lab}{the node label of the coalescent node}
  \item{Re.lab}{the node label of the reassortment node}
  \item{node.lab1}{a node label}
  \item{node.lab2}{another node label}
  \item{nwktree}{the tree string in newick format}
  \item{parent.numset}{a set of numbers indicating the candicate ancestors}
  \item{desc.numset}{a set of numbers indicating the candicate descendants}
  \item{nodecol}{the specified type/color (i.e. 'r' for samples, 'g' for branching nodes, '#Ha1' and '#Ha2' for reassortment nodes) of the requested nodes}
  \item{incongr.nodes}{the discording coalescent nodes between trees}
  \item{reversed.nodes}{the reversed coalescent nodes between trees}
}}
}
\value{
A string for the node/tip

A string for the node/tip

A string for the node/tip

A vector of characters

An vector of characters

logical

An string

A string

An vector of characters

An vector of characters
}
\description{
Track/get components from a tree
}
